test(Patient): :test_tube: tests cover all the class and integrate a new function to remove a Prescription
Note on the remove function : We couldn't actually test the remove function because it is strongly
link to the interface, this is an issue that need to be solved in the refactor branch.
To be able to check if remove works we need to overload the function to find the right one to delete.